long改bigdecimal

+++ src/main/java/com/ogc/standard/ao/IAlipayAO.java
@@ -8,6 +8,8 @@
  */
 package com.ogc.standard.ao;
 
+import java.math.BigDecimal;
+
 /** 
  * @author: haiqingzheng 
  * @since: 2017年1月11日 下午8:55:34 
@@ -17,7 +19,7 @@ public interface IAlipayAO {
     // 支付宝APP支付，第一步：获取签名后的订单信息。
     public Object getSignedOrder(String applyUser, String toUser,
             String payGroup, String refNo, String bizType, String bizNote,
-            Long transAmount, String backUrl);
+            BigDecimal transAmount, String backUrl);
 
     public void doCallbackAPP(String result);
